♻️ Add Vyper Support in `foundry.toml` File (#252)
### 🕓 Changelog

The Foundry PR https://github.com/foundry-rs/foundry/pull/8061 added the
`skip` key to `foundry.toml`, which can be used to configure skipping
compilation of non-compilable files. We use this new feature to only
compile 🐍 snekmate mock contracts. In addition, the Foundry PR
https://github.com/foundry-rs/foundry/pull/8086 added a Vyper
configuration section `vyper` to `foundry.toml` which allows to
configure the attributes `optimize` and `path`. We set the `optimize`
mode to the compiler default value `gas`.

---------

Signed-off-by: Pascal Marco Caversaccio <pascal.caversaccio@hotmail.ch>