⚡️ Optimise `erc1155` `NoBackdoor` `halmos` Test (#255)
### 🕓 Changelog

This PR uncomments the `erc721` and `erc1155` `NoBackdoor` `halmos`
tests and optimises the `erc1155` `NoBackdoor` `halmos` test. In
addition, we remove the `test-parallel` flag in the `halmos`
configuration, because if both `test-parallel` and `solver-parallel` are
enabled, the (peak) number of parallel solver processes could be too
high, which could affect performance. The `erc721` `halmos` tests take
substantial amount of time as part of the CI. Henceforth, we run the
`halmos` CI pipeline each day at _03:30 a.m._ (= "nightly" tests) as
scheduled `cron` job instead of integrating it into the normal `push`
and `pull_request` pipeline.

---------

Signed-off-by: Pascal Marco Caversaccio <pascal.caversaccio@hotmail.ch>
Co-authored-by: Pascal Marco Caversaccio <pascal.caversaccio@hotmail.ch>