👷 Simplify `halmos.toml` by Switching to Bundled Yices 2 SMT Solver (#326)
### 🕓 Changelog

As of [`a16z/halmos#525`](https://github.com/a16z/halmos/pull/525),
`halmos` now ships with a `yices-smt2`
[`2.6.4`](https://github.com/SRI-CSL/yices2/releases/tag/Yices-2.6.4)
binary, so there's no longer a need to install the [Yices 2 SMT
solver](https://github.com/SRI-CSL/yices2) separately in the CI
environment. This commit updates the `halmos.toml` configuration file to use
the higher-level `solver = "yices-2.6.4"` setting instead of the
lower-level `solver-command` field. Functionally, this change is a no-op
(i.e. the same solver is used) but it's now sourced from the bundled
binary instead of a release tarball. This simplifies the configuration
and aligns with the upstream best practices.

---------

Signed-off-by: Pascal Marco Caversaccio <pascal.caversaccio@hotmail.ch>
Co-authored-by: Pascal Marco Caversaccio <pascal.caversaccio@hotmail.ch>