🥢 Set `_skipRoundTripShares` Flag in `ERC4626VaultTest` (#336)
### 🕓 Changelog

The test library
[`erc4626-tests`](https://github.com/a16z/erc4626-tests) introduced the
`_skipRoundTripShares` flag in PR
[a16z/erc4626-tests#15](https://github.com/a16z/erc4626-tests/pull/15).
When set to `true`, it skips the shares inequality assertions in
round‑trip tests. The default value is `false`, but since we favour
explicitness, we explicitly set `_skipRoundTripShares = false` in the
`ERC4626VaultTest` test `setUp` function. Furthermore, this commit pins
[`forge-std`](https://github.com/foundry-rs/forge-std) to commit
[`369dd01c808c255765b074c12aa420da33947a65`](https://github.com/foundry-rs/forge-std/commit/369dd01c808c255765b074c12aa420da33947a65)
to temporarily prevent CI test failures in the nightly
[`halmos`](https://github.com/a16z/halmos)-based tests (see
[foundry-rs/forge-std#705](https://github.com/foundry-rs/forge-std/issues/705)).

---------

Signed-off-by: Pascal Marco Caversaccio <pascal.caversaccio@hotmail.ch>